Silence extraneous output from jlist
Calling pm2 jlist when the pm2 service is not running triggers pm2
to start the daemon process and emit banner and status text. This is not
parseable json and breaks capistrano-pm2's app_status method when ran on an
instance where the daemon is not started.

This patch adds the -s silent running argument introduced into pm2 version
2.4.0.